Add ability to archive invoices on our side
