+++ app/views/admin/publishers/invoices/edit.html.slim
