+++ app/views/admin/publishers/invoices/finalize.html.slim
