+++ app/views/admin/publishers/invoices/new.html.slim
