+++ app/views/admin/publishers/invoices/show.html.slim
@@ -1,10 +1,7 @@
 
 .d-flex
-  h2.mt-1= @invoice.human_date + " Invoice"
-  - if @invoice.in_progress?
-    .ml-3
-      = link_to "Generate Manual Payout", admin_publisher_generate_manual_payout_path(id: @invoice.publisher), class: 'btn btn-info'
-  - elsif @invoice.pending?
+  h3.mt-1= @invoice.human_date + " Invoice"
+  - if @invoice.pending?
     .ml-3
       = link_to "Edit", edit_admin_publisher_invoice_path(@invoice), class: 'btn btn-primary'
     .ml-3
@@ -57,6 +54,7 @@ table.display.table.table-bordered.table-striped
     th Uploaded By
     th Archived
     th Created At
+    th
   tbody
   - @invoice.invoice_files.each do |invoice_file|
     tr
@@ -64,6 +62,9 @@ table.display.table.table-bordered.table-striped
         = link_to invoice_file.file.filename, url_for(invoice_file.file)
       td= invoice_file.uploaded_by
       td
-       - if invoice_file.archived
-        = fa_icon "archive"
+        - if invoice_file.archived
+          = fa_icon "archive"
       td= invoice_file.created_at
+      td
+        - text = invoice_file.archived? ? "Restore file" : "Archive File"
+        = link_to text, admin_publisher_invoice_archive_file_path(invoice_id: @invoice.id, invoice_file_id: invoice_file.id),  method: :post, class: ''
